Move inmem store to own file
